// react-app/src/types.ts
// types.ts
export type Appointment = {
    id: number;
    nextAppointmentDate: string;
    contractedSales: number;
    currentContractCount: number;
    companyName: string;
    companyNameKana: string;
    capital: number;
    employeesCount: number;
    appointmentDepartment:string;
    contactPersonName:string;
    contactPersonKana:string;
    url:string;
    place:string;
    appointmentData:string;
    history:string;
    goal:number;
};